4-byte moves i guess
add the dummy tm compatabilities
add the possibility to set all the dex flags for flexing purposes
update the learnsets
increase the max moves/learnset to 30

+++ src/data/pokemon/level_up_learnset_pointers.h
@@ -5,7 +5,7 @@
 #ifndef POKERUBY_LEVEL_UP_LEARNSET_POINTERS_H
 #define POKERUBY_LEVEL_UP_LEARNSET_POINTERS_H
 
-const u16 *gLevelUpLearnsets[NUM_SPECIES] =
+const u32 *gLevelUpLearnsets[NUM_SPECIES] =
 {
     [SPECIES_NONE] = sBulbasaurLevelUpLearnset,
     [SPECIES_BULBASAUR] = sBulbasaurLevelUpLearnset,
